+++ packages/vrender-components/CHANGELOG.md
@@ -1,6 +1,19 @@
 # Change Log - @visactor/vrender-components
 
-This log was last generated on Wed, 04 Sep 2024 12:52:31 GMT and should not be manually modified.
+This log was last generated on Sat, 07 Sep 2024 09:16:33 GMT and should not be manually modified.
+
+## 0.20.3
+Sat, 07 Sep 2024 09:16:33 GMT
+
+### Updates
+
+- feat: add `alignOffset` in arc-label, `line2MinLength` support customized callback
+
+
+- feat: support axis `breaks` for line axis
+- feat: support `autoWrap` in line axis label
+- feat: label component supports disable specific state animation
+- fix: tag padding not work when label is rich text. fix@VisActor/VChart#3151
 
 ## 0.20.2
 Wed, 04 Sep 2024 12:52:31 GMT